+++ Classes/Model/FrontEndUser.php
@@ -11,22 +11,4 @@ use OliverKlee\Oelib\Model\FrontEndUser as OelibFrontEndUser;
  */
 class FrontEndUser extends OelibFrontEndUser
 {
-    /**
-     * Gets the registration record for which this user is related to as "additional registered person".
-     */
-    public function getRegistration(): ?Registration
-    {
-        /** @var Registration|null $registration */
-        $registration = $this->getAsModel('tx_seminars_registration');
-
-        return $registration;
-    }
-
-    /**
-     * Sets the registration record for which this user is related to as "additional registered person".
-     */
-    public function setRegistration(?Registration $registration = null): void
-    {
-        $this->set('tx_seminars_registration', $registration);
-    }
 }